Cerulean Initial Commit!
Welcome to version control, Cerulean!

Implements the first round of this tool.  Defines a number of objects:
  - ETL_Process: a JSON representation of a PHP-ETL process
  - ETL_DB: a JSON representation of a PHP-ETL database (source/destination for extracts/loads)
  - ETL_Record: a place to stash individual records (in JSON) between processes
  - ETL_RecordType:  a type of the above; defines the Schema
  - ETL_Admin: ModelAdmin extensions to manage the above

We also define a set of our own Extractors, Transformers and Loaders; some of these
will hopefully be folded into PHP-ETL directly, but until they are, we'll store them
locally.

Includes the necessary composer and configs to install the pre-reqs and make it all
work... hopefully.  More commits to come to make sure everything is actually fitting together