MARC Extractor!
Pretty specialized at this point, but does the job better than trying
to treat MARC as XML.... it's really not, even when it is.

1. id_path defines which tag/marcspec path gives you a meaningful
ID for the record; defaults to 001, and will fallback to filename +
record number if nothing is returned.
2. holdings_tag defines where the item information is stored.  If this
is given, then item information is put into the row; else it's biblio
info
3. columns are marcspec paths for biblios, subfield codes for items
4. separator, in case of multiple returns