RESTful services support part 1
Creates a manager for RESTful services, modelled after Databases
in PHP-ETL.  Relies on Guzzle to do the actual connection work.

Adds ETL_REST, which is like ETL_DB, for tracking sources in SS.

Adds ETL_REST_Schema.json for generating the form for the JSON
in the Configuration section.

Updates RunETLProcesses to loop through RESTful sources and add
to manager.

Also fixes:
 - required package for MARC Extractor missing from composer.json
 - ETL_DB loop in RunETLProcesses not grabbing right data